pyproject.toml
[project]
name = "knowledge-platform"
version = "0.1.0"
description = "RAG Knowledge Platform"
authors = [
    {name = "mizhexiaoxiao", email = "mizhexiaoxiao@gmail.com"},
]
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.135.1",
    "tortoise-orm==0.23.0",
    "pydantic==2.12.5",
    "email-validator==2.3.0",
    "passlib==1.7.4",
    "pyjwt==2.11.0",
    "black==25.1.0",
    "isort==5.13.2",
    "ruff==0.9.1",
    "loguru==0.7.3",
    "pydantic-settings==2.13.1",
    "argon2-cffi==23.1.0",
    "pydantic-core==2.41.5",
    "annotated-types==0.7.0",
    "setuptools==69.5.1",
    "uvicorn==0.41.0",
    "h11==0.14.0",
    "aerich==0.8.1",
    "aiosqlite==0.20.0",
    "anyio==4.10.0",
    "argon2-cffi-bindings==21.2.0",
    "asyncclick==8.1.8",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "click==8.3.1",
    "dictdiffer==0.9.0",
    "dnspython==2.8.0",
    "fastapi-cli==0.0.7",
    "httpcore==1.0.5",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.7",
    "iso8601==2.1.0",
    "jinja2==3.1.4",
    "markdown-it-py==4.0.0",
    "markupsafe==2.1.5",
    "mdurl==0.1.2",
    "mypy-extensions==1.0.0",
    "orjson==3.10.6",
    "packaging==26.0",
    "pathspec==0.12.1",
    "platformdirs==4.9.2",
    "pycparser==2.21",
    "pygments==2.18.0",
    "pypika-tortoise==0.3.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.22",
    "pytz==2024.1",
    "pyyaml==6.0.3",
    "rich==14.3.3",
    "rich-toolkit==0.13.2",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.52.1",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "ujson==5.10.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.4",
    "websockets==14.1",
    "asyncpg>=0.31.0",
    "async>=0.6.2",
    "rq>=2.8.0",
    "rq-dashboard-fast>=0.9.2",
    "pytest>=9.0.3",
    "requests>=2.33.1",
    "openai>=2.32.0",
    "lxml>=6.1.0",
    "aiohttp>=3.13.5",
]

[tool.black]
line-length = 120
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 120
lint.extend-select = []
lint.ignore = [
    "F403",
    "F405",
]

[tool.aerich]
tortoise_orm = "config.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
